# swagger.yaml
swagger: '2.0'
info:
  title: Todo API
  description: Manage todos
  version: 1.0.0
host: localhost:9999
schemes:
  - http
consumes:
  - application/json
produces:
  - application/json
definitions:
  Todo:
    description: A task that has to be done
    type: object
    properties:
      id:
        type: string
        readOnly: true
      title:
        type: string
        minLength: 1
      done:
        type: boolean
        readOnly: true
    required:
      - id
      - title
      - done
  Todos:
    description: A list of todos
    type: array
    items:
      $ref: '#/definitions/Todo'
paths:
  /todos:
    get:
      operationId: getTodos
      responses:
        200:
          description: Returns all todos
          schema:
            $ref: '#/definitions/Todos'
    post:
      operationId: createTodo
      parameters:
        - name: todo
          in: body
          required: true
          schema:
            $ref: '#/definitions/Todo'
      responses:
        200:
          description: The newly created todo
          schema:
            $ref: '#/definitions/Todo'
        400:
          description: Invalid data
  /todos/{todoId}:
    parameters:
      - name: todoId
        in: path
        type: string
        required: true
    get:
      operationId: getTodo
      responses:
        200:
          description: The todo with the given todo id
          schema:
            $ref: '#/definitions/Todo'
        404:
          description: Unknown todo id
    put:
      operationId: updateTodo
      parameters:
        - name: update
          in: body
          required: true
          schema:
            $ref: '#/definitions/Todo'
      responses:
        200:
          description: Successfully updated the todo
        400:
          description: Invalid data
        404:
          description: Unknown todo id
    delete:
      operationId: deleteTodo
      responses:
        200:
          description: Successfully deleted the todo
        404:
          description: Unknown todo id
  /todos/{todoId}/setdone:
    parameters:
      - name: todoId
        in: path
        type: string
        required: true
    put:
      operationId: setDone
      responses:
        200:
          description: Successfully set the todo as done
        404:
          description: Unknown todo id